ch3-drf/requirements.txt
asgiref==3.8.1
Django==4.2
django-seed==0.3.1
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
Faker==33.3.1
psycopg2==2.9.10
PyJWT==2.10.1
python-dateutil==2.9.0.post0
six==1.17.0
sqlparse==0.5.3
toposort==1.10
typing_extensions==4.12.2
tzdata==2025.1
